FIX #13313

+++ htdocs/core/tpl/objectline_create.tpl.php
@@ -761,7 +761,7 @@ if (!empty($usemargins) && $user->rights->margins->creer)
 
 		/* To process customer price per quantity (CUSTOMER_PRICE_PER_QTY works only if combo product is not an ajax after x key pressed) */
 		var pbq = parseInt($('option:selected', this).attr('data-pbq'));
-		var pbqup = parseInt($('option:selected', this).attr('data-pbqup'));
+		var pbqup = parseFloat($('option:selected', this).attr('data-pbqup'));
 		var pbqbase = $('option:selected', this).attr('data-pbqbase');
 		var pbqqty = parseFloat($('option:selected', this).attr('data-pbqqty'));
 		var pbqpercent = parseFloat($('option:selected', this).attr('data-pbqpercent'));
